Factor out common cloud storage client functionality
